Added swap method

+++ backend/src/main/kotlin/res/ActiveTasksResource.kt
@@ -6,7 +6,6 @@ import io.ktor.request.receive
 import io.ktor.response.respond
 import io.ktor.routing.*
 import model.ActiveTaskDTO
-import res.ResourcesUtil.withTaskId
 import res.ResourcesUtil.withUserId
 import services.ActiveTasksService
 
@@ -29,22 +28,50 @@ fun Route.activeTasks(activeTasksService: ActiveTasksService) {
 
         put("/{id}") {
             withUserId(call) { userId ->
-                withTaskId(call) { taskId ->
-                    val task = call.receive<ActiveTaskDTO>()
-                    val updated = activeTasksService.updateTask(taskId, task, userId)
-                    if (updated == null) call.respond(HttpStatusCode.NotFound)
-                    else call.respond(HttpStatusCode.NoContent)
+                val taskId = call.parameters["id"]?.toInt()
+                if (taskId == null) {
+                    call.respond(HttpStatusCode.BadRequest)
+                    return@withUserId
                 }
+                val task = call.receive<ActiveTaskDTO>()
+                val updated = activeTasksService.updateTask(taskId, task, userId)
+                if (updated == null) call.respond(HttpStatusCode.NotFound)
+                else call.respond(HttpStatusCode.NoContent)
             }
         }
 
         delete("/{id}") {
             withUserId(call) { userId ->
-                withTaskId(call) { taskId ->
-                    val removed = activeTasksService.deleteTask(taskId, userId)
-                    if (removed) call.respond(HttpStatusCode.NoContent)
-                    else call.respond(HttpStatusCode.NotFound)
+                val taskId = call.parameters["id"]?.toInt()
+                if (taskId == null) {
+                    call.respond(HttpStatusCode.BadRequest)
+                    return@withUserId
                 }
+                val removed = activeTasksService.deleteTask(taskId, userId)
+                if (removed) call.respond(HttpStatusCode.NoContent)
+                else call.respond(HttpStatusCode.NotFound)
+            }
+        }
+
+        post("/swap/{id1}/{id2}") {
+            withUserId(call) { userId ->
+                val taskId1 = call.parameters["id1"]?.toInt()
+                val taskId2 = call.parameters["id2"]?.toInt()
+                if (taskId1 == null || taskId2 == null) {
+                    call.respond(HttpStatusCode.BadRequest)
+                    return@withUserId
+                }
+
+                val task1 = activeTasksService.getTask(taskId1, userId)
+                val task2 = activeTasksService.getTask(taskId2, userId)
+                if (task1 == null || task2 == null) {
+                    call.respond(HttpStatusCode.NotFound)
+                    return@withUserId
+                }
+
+                activeTasksService.updateQueueNumber(taskId1, task2.inQueue!!, userId)
+                activeTasksService.updateQueueNumber(taskId2, task1.inQueue!!, userId)
+                call.respond(HttpStatusCode.NoContent)
             }
         }
     }
